Restore a set of empty parenthesis in template
If the template calls JavaScript functions, then the parenthesis
will be interpreted as an Elisp macro and attempted to execute. However,
since it is the empty string, there'll be no content, but the
parenthesis is removed, making the JavaScript evaluate wrongly.
(This is currently still a problem with JavaScript taking parameters,
but that can more easily be avoided in the templates).