Write changed files only
Test if the contents in buffer has changed compared to the one on disk
and only commit it if there are differences, to avoid touching the disk
only to rewrite the same bytes.

Disadvantage of this approach is that the entire blog content still has
to be generated in each run. A better approach would perhaps be to hash
the source org-mode contents (together with the headers) and store it in
a cache somewhere.