One sub-headline level, but configurable in org-file
Default number of headlines is set to 2, which means that each post can
have one title and one level of subsections. Third level indentation is
then considered to be a bulleted list.

This setting can be changed in the org-file by using an #+OPTIONS line
and setting the "H" export option (currently this is the only one that
is supported).